+++ code/train_ISTL_UCSDPed1&2.py
@@ -303,10 +303,10 @@ for p in params:
 					file=sys.stderr)
 
 		q['training_rec_errors'] = {'1st iteration': {
-											'mean': train_rec_error.mean(),
-											'std': train_rec_error.std(),
-											'min': train_rec_error.min(),
-											'max': train_rec_error.max()
+											'mean': float(train_rec_error.mean()),
+											'std': float(train_rec_error.std()),
+											'min': float(train_rec_error.min()),
+											'max': float(train_rec_error.max())
 										}}
 
 		#	Test on UCSD Ped 1
@@ -501,10 +501,10 @@ for p in params:
 						file=sys.stderr)
 
 			q['training_rec_errors']['2nd iteration'] = {
-											'mean': train_rec_error.mean(),
-											'std': train_rec_error.std(),
-											'min': train_rec_error.min(),
-											'max': train_rec_error.max()
+											'mean': float(train_rec_error.mean()),
+											'std': float(train_rec_error.std()),
+											'min': float(train_rec_error.min()),
+											'max': float(train_rec_error.max())
 										}
 
 			### Evaluation of test set
@@ -694,10 +694,10 @@ for p in params:
 						file=sys.stderr)
 
 			q['training_rec_errors']['3rd iteration'] = {
-											'mean': train_rec_error.mean(),
-											'std': train_rec_error.std(),
-											'min': train_rec_error.min(),
-											'max': train_rec_error.max()
+											'mean': float(train_rec_error.mean()),
+											'std': float(train_rec_error.std()),
+											'min': float(train_rec_error.min()),
+											'max': float(train_rec_error.max())
 										}
 
 		else:
